Resolve "Update CoW tutorial"

Closes #2116

See merge request nomad-lab/nomad-FAIR!2066

+++ examples/data/cow_tutorial/nomad-countries/src/nomad_countries/schema_packages/country.py
@@ -9,7 +9,7 @@ m_package = Package(name='Countries of the World')
 
 
 class Timeseries(MSection):
-    year = Quantity(type=np.int, shape=['*'])
+    year = Quantity(type=np.int32, shape=['*'])
     value = Quantity(type=np.float64, shape=['*'])
